Styled-main-page
