+++ Bit-Final-Project/interview-reports/src/components/candidate-reports/CandidateReport.jsx
@@ -1,5 +1,5 @@
 import BasicInfo from "./BasicInfo"
-import Header from "./Header"
+
 import './CandidateReport.css';
 import Table from "./Table";
 
@@ -8,7 +8,6 @@ const CandidateReport = () => {
     return (
         <>
         <div className="container-liquid info-container">
-            <Header />
             <BasicInfo />
         </div>
         <div className="container table">
