+++ Bit-Final-Project/interview-reports/src/components/card/card.css
@@ -1,9 +1,17 @@
-.card-image{
-    width: 200px;
+.img-container{
+    width: 40%;
 }
 .glavni{
-    width: 30%;
-    -webkit-box-shadow: 0px 0px 48px 5px rgba(0,0,0,1);
--moz-box-shadow: 0px 0px 48px 5px rgba(0,0,0,1);
-box-shadow: 0px 0px 48px 5px rgba(0,0,0,1);
+    -webkit-box-shadow: 4px 4px 18px 2px rgba(0,0,0,0.38);
+-moz-box-shadow: 4px 4px 18px 2px rgba(0,0,0,0.38);
+box-shadow: 4px 4px 18px 2px rgba(0,0,0,0.38);
+transition: box-shadow .4s;
+}
+
+.glavni:hover {
+    box-shadow: 0 0 11px rgba(33,33,33,.15); 
+  }
+
+.paragraph{
+    word-break: break-all;
 }
