+++ Bit-Final-Project/interview-reports/src/components/main_page/main.jsx
@@ -11,7 +11,7 @@ export const Main = () => {
   }, []);
 
   return (
-    <div className="container p-5 d-flex flex-wrap justify-content-between bg-light">
+    <div className="container d-flex flex-wrap justify-content-center mt-5">
       {Object.values(user).map((value, index) => {
         return <Card data={value} key={index} />;
       })}
